+++ modules/carto/src/layers/carto-layer.js
@@ -1,14 +1,15 @@
 import {CompositeLayer} from '@deck.gl/core';
 import {MVTLayer} from '@deck.gl/geo-layers';
 import {GeoJsonLayer} from '@deck.gl/layers';
-import {getMap, CONNECTIONS, MAP_TYPES, MODE_TYPES, FORMATS, getTileJSON} from '../api/maps-api-client';
-import {getMapsVersion} from '../config';
+import {getMap} from '../api/maps-api-client';
 
 const defaultProps = {
   // (String, required): data resource to load. table name, sql query or tileset name.
   data: null,
   // carto credentials, set to null to read from default
   credentials: null,
+  // mode should carto or carto-cloud-native
+  mode: null,
   // (String {bigquery, snowflake,redshift, postgres}, required)
   provider: null,
   // (String, required): connection name at CARTO platform
@@ -23,15 +24,22 @@ const defaultProps = {
   onDataError: {type: 'function', value: null, compare: false, optional: true}
 };
 
+function getSublayerFromMapFormat(format) {
+  if (format === 'tilejson') {
+    return MVTLayer;
+  }
+  else if (format === 'geojson') {
+    return GeoJsonLayer;
+  }
+
+  return null;
+}
 export default class CartoLayer extends CompositeLayer {
   initializeState() {
     this.state = {
       data: null,
       SubLayer: null
     };
-
-    // const {mode, type} = this.props;
-    // checkAllowedProps({mode, type});
   }
 
   get isLoaded() {
@@ -53,31 +61,23 @@ export default class CartoLayer extends CompositeLayer {
 
   async _updateData() {
     try {
+    
       const {mode, provider, type, data: source, connection, credentials, format} = this.props;
 
-      let data;
-      let SubLayer;
+      let data, mapFormat;
 
-      if (mode === MODE_TYPES.CARTO)  {
-        if (type === MAP_TYPES.SQL || type == MAP_TYPES.TILESET) {
-          data = await this.updateSQLTileJSON();
-        }
-
-        if (type === MAP_TYPES.TILESET) {
-          data = await this.updateBQTilerTileJSON();
-        }
-
-        SubLayer = this.state.SubLayer || this.props.subLayer || getSublayerFromMapFormat(FORMATS.TILEJSON);
+      if (mode === mode.CARTO_CLOUD_NATIVE) {
+        [data, mapFormat] = await getMap({provider, type, source, connection, credentials, format});
+      } else {
+        data = await this.updateTileJSON();
+        mapFormat = FORMATS.TILEJSON
       }
 
-      if (mode === MODE_TYPES.CARTO_CLOUD_NATIVE) {
-        const [mapData, mapFormat] = await getMap({provider, type, source, connection, credentials, format});
-        data = mapData;
-        SubLayer = this.state.SubLayer || this.props.subLayer || getSublayerFromMapFormat(mapFormat);
-      }
+      const SubLayer = this.state.SubLayer || this.props.subLayer || getSublayerFromMapFormat(mapFormat);
 
       this.setState({SubLayer, data});
       this.props.onDataLoad(data);
+
     } catch (err) {
       if (this.props.onDataError) {
         this.props.onDataError(err);
@@ -87,56 +87,8 @@ export default class CartoLayer extends CompositeLayer {
     }
   }
 
-   async updateSQLTileJSON() {
-    const {data, bufferSize, tileExtent, credentials} = this.props;
-    const version = getMapsVersion(credentials);
-    const isSQL = data.search(' ') > -1;
-
-    switch (version) {
-      case 'v1':
-        const sql = isSQL ? data : `SELECT * FROM ${data}`;
-
-        // Map config v1
-        const mapConfig = {
-          version: '1.3.1',
-          buffersize: {
-            mvt: bufferSize
-          },
-          layers: [
-            {
-              type: 'mapnik',
-              options: {
-                sql: sql.trim(),
-                vector_extent: tileExtent
-              }
-            }
-          ]
-        };
-
-        return await getTileJSON({mapConfig, credentials});
-
-      case 'v2':
-        return await getTileJSON({
-          connection: CONNECTIONS.CARTO,
-          source: data,
-          type: isSQL ? MAP_TYPES.SQL : MAP_TYPES.TABLE,
-          credentials
-        });
-
-      default:
-        throw new Error(`Cannot build MapConfig for unmatching version ${version}`);
-    }
-  }
-
-  async updateBQTilerTileJSON() {
-    const {credentials, data} = this.props;
-
-    return await getTileJSON({
-      connection: CONNECTIONS.BIGQUERY,
-      type: MAP_TYPES.TILESET,
-      source: data,
-      credentials
-    });
+  async updateTileJSON() {
+    throw new Error('You must use one of the specific carto layers: BQ or SQL type');
   }
 
   renderLayers() {
@@ -161,14 +113,3 @@ export default class CartoLayer extends CompositeLayer {
 
 CartoLayer.layerName = 'CartoLayer';
 CartoLayer.defaultProps = defaultProps;
-
-function getSublayerFromMapFormat(format) {
-  switch (format) {
-    case FORMATS.TILEJSON:
-      return MVTLayer;
-    case FORMATS.GEOJSON:
-      return GeoJsonLayer;
-    default:
-      return null;
-  }
-}
